feat: add tools start logic to dtm start command
