+++ internal/pkg/start/tool/helm.go
@@ -1,0 +1,26 @@
+package tool
+
+import (
+	"fmt"
+	"os/exec"
+)
+
+var toolHelm = tool{
+	Name: "Helm",
+
+	Exists: func() bool {
+		_, err := exec.LookPath("helm")
+		return err == nil
+	},
+
+	Install: func() error {
+		if !confirm("Helm") {
+			return fmt.Errorf("user cancelled")
+		}
+
+		if err := execCommand([]string{"brew", "install", "helm"}); err != nil {
+			return err
+		}
+		return nil
+	},
+}
